chore: split verbs/mod -> impl_maths
